Fixed bot

+++ app.py
@@ -224,19 +224,19 @@ def home():
             except:
                 value = "Foo"
             if message in ["Hi", "hi", "Hello", "hello", "Hey", "hey", "0"]:
-                # resp = get_profile(phone=phone[2:])
-                # if resp['ReplyCode'] != 0:
-                #     returnMessage = "Please get yourself registered first!\n\nRegister here\nhttp://tsite.paapos.in/"
-                # else:
-                db_operations.delete_one({'_id': int(phone)})
-                new_user = {
-                    '_id': int(phone),
-                    "CT": "C",
-                    'cid': resp['Cid']
-                }
-                db_operations.insert_one(new_user)
-                user = db_operations.find_one({'_id': int(phone)})
-                returnMessage = "Welcome to paapos, your preferred delivery partner\n\nTo Book Same Day Order Reply 1\nTo Book Courier Reply 2\nTo Track Your Order Reply 3\nTo cancel your order reply 4\nTo connect with our Executive reply 5\nFor feedback reply 6\n\nFor the main menu, type *0*\nFor the previous menu, type *9*"
+                resp = get_profile(phone=phone[2:])
+                if resp['ReplyCode'] != 0:
+                    returnMessage = "Please get yourself registered first!\n\nRegister here\nhttp://tsite.paapos.in/"
+                else:
+                    db_operations.delete_one({'_id': int(phone)})
+                    new_user = {
+                        '_id': int(phone),
+                        "CT": "C",
+                        'cid': resp['Cid']
+                    }
+                    db_operations.insert_one(new_user)
+                    user = db_operations.find_one({'_id': int(phone)})
+                    returnMessage = "Welcome to paapos, your preferred delivery partner\n\nTo Book Same Day Order Reply 1\nTo Book Courier Reply 2\nTo Track Your Order Reply 3\nTo cancel your order reply 4\nTo connect with our Executive reply 5\nFor feedback reply 6\n\nFor the main menu, type *0*\nFor the previous menu, type *9*"
                 updated_user = {"$set": {'returnMessage' : returnMessage}}
                 db_operations.update_one(user, updated_user)
                 return send_message(message=returnMessage, phone=phone)
